organisation and semantic changes + add error handling in axios :cyclone:
