--- src/Results.js
@@ -1,19 +1,0 @@
-// Results
-
-const Results = (props) => {
-    return (
-        <div className="user-horoscopeReading">
-            <p>{props.current_date}</p>
-            <p>{props.description}</p>
-            <p>{props.compatibility}</p>
-            <p>{props.color}</p>
-            <p>{props.lucky_number}</p>
-            <p>{props.lucky_time}</p>
-            <p>{props.mood}</p>
-        </div>
-
-        
-    )
-}
-
-export default Results;
